refactor(cockpit): extract file stat module
